Split out speechManager into speech.manager module and priority constants into  speech.priorities module.
